13.184(2) Deleting a Song from the Storage/Database
Delete a song from the firebase database.
The song is in the upload array. It will need to be removed from the list in the manage component. We are in a child component. We don't have access to the array. We only have access to a single song. To solve this we create a function in the manage component.